Rename inference-engine/3rd_party/marian-nmt
