+++ .gitmodules
@@ -10,13 +10,31 @@
 	path = inference-engine/bergamot-translator-tests
 	url = https://github.com/browsermt/bergamot-translator-tests
 
+# This is the same dependency and repository as `3rd_party/browsermt-marian-dev` below.
+#
+# When forking `inference-engine` into to this project, I made an earnest attempt to utilize the preexisting
+# `3rd_party/browsermt-marian-dev` submodule within `inference-engine`. Unfortunately, I ran into several roadblocks:
+#
+#   1) I cannot directly add `3rd_party/browsermt-marian-dev` as a cmake subdirectory because cmake is aware that
+#      this path is not a subdirectory of the `inference-engine` project root.
+#
+#   2) Symbolic links do not appear to work for git submodule direcotires the way that they do for regular directories.
+#      Even if the symbolic link had linked correctly, it may have still failed due to the considerations of 1).
+#
+#   3) I tried using cmake to copy the files from `3rd_party/browsermt-marian-dev` into `inference-engine/3rd_party/browsermt-marian-dev`
+#      at build time, which would ensure that there is no duplicate reference to the URL in this file, however the upstream dependency itself
+#      has hard-coded cmake expectations that the `.git` directory is only one level up, which appears to work correctly for the way git submodules
+#      are configured, but does not work if the files are copied over to a regular directory deeper in the repository.
+#
+# It may be possible to remove `3rd_party/browsermt-marian-dev` to instead use `inference-engine/3rd-party/browsermt-marian-dev` everywhere
+# within this repository, but I will leave that for a future commit if there is a need to do so.
+[submodule "inference-engine/3rd_party/browsermt-marian-dev"]
+	path = inference-engine/3rd_party/browsermt-marian-dev
+	url = https://github.com/browsermt/marian-dev
+
 [submodule "inference-engine/3rd_party/pybind11"]
 	path = inference-engine/3rd_party/pybind11
 	url = https://github.com/pybind/pybind11.git
-
-[submodule "inference-engine/3rd_party/marian-dev"]
-	path = inference-engine/3rd_party/marian-dev
-	url = https://github.com/browsermt/marian-dev
 
 [submodule "inference-engine/3rd_party/ssplit-cpp"]
 	path = inference-engine/3rd_party/ssplit-cpp
