+++ inference-engine/src/translator/parser.h
@@ -4,7 +4,7 @@
 #include <fstream>
 #include <sstream>
 
-#include "3rd_party/marian-dev/src/3rd_party/CLI/CLI.hpp"
+#include "3rd_party/browsermt-marian-dev/src/3rd_party/CLI/CLI.hpp"
 #include "3rd_party/yaml-cpp/yaml.h"
 #include "common/build_info.h"
 #include "common/config_parser.h"
